Move in-memory tests into keychain
